requirements.txt
cdt==0.6.0
einops==0.8.0
gcastle==1.0.3
networkx==2.8.8
numpy==1.24.1
opt-einsum==3.3.0
pandas==2.0.3
pygam==0.8.0
PyYAML==6.0.2
tensorly==0.8.1
tensorly-torch==0.5.0
torch==2.4.0
wandb==0.19.6
gpustat==1.1.1
openai==1.62.0
langchain-openai==0.1.25
python-dotenv==1.0.1
bnlearn==0.10.2
causal-learn==0.1.4.1
transformers==4.51.3